[Update] Move Marker Design Func 3 (Clinic, Hospital Code Split)
