+++ app/src/main/java/com/golddog/mask_location/base/BaseActivity.kt
@@ -7,11 +7,11 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.BindingAdapter
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.FragmentManager
 
 abstract class BaseActivity<T : ViewDataBinding> : AppCompatActivity() {
 
     protected abstract val layoutResourceId: Int
-
     protected lateinit var binding: T
 
     override fun onCreate(savedInstanceState: Bundle?) {
